Fixed redis-cli readLine loop to correctly handle EOF.
When using the shutdown command with redis-cli the server saves the database and, if successful, silently closes the connection. The redis-cli tool did not correcty handle this EOF case in its readLine loop and was therefore infinitely looping - and eating 100% of the CPU - while waiting for some data which would never come.

--- Replay Metadata ---
{
  "original_author": {
    "name": "Luc Heinrich",
    "email": "luc@honk-honk.com"
  },
  "original_committer": {
    "name": "Luc Heinrich",
    "email": "luc@honk-honk.com"
  },
  "original_dates": {
    "author": "2009-03-23T11:43:16+00:00",
    "commit": "2009-03-23T11:43:16+00:00"
  },
  "original_commit": "b91f03a4d14c166bb8ead41e5b556b1080ccd559",
  "assigned_team_member": {
    "name": "Jordan Smith",
    "email": "jordan-smith@gcp-test-qa.com",
    "id": "jordan-smith"
  },
  "replay_timestamp": "2025-10-21T18:12:29.292167+00:00"
}
--- End Replay Metadata ---